Integrate astenc v2.1, new sse2neon, more simd sources, add decode filter tests/sets encoder now
added astcenc v2.1.  Removed some changes out of astenc_image, but still needed USE_2DARRAY support.  Remove 1/2 channel optimizations for now in astcenc_pickbestendpointformat.  Lots of changes due to astcenc-side vector change from xyzw to rgba.

fix float4a issues, and complete the api, fixed lengthSquared call
sse2 has arm64 support for permute, and newer neon instructions.
add AVX and SSE4.1 ops for round/floor/ceil and fp16 <-> fp32 conversion, this re-enables fp16 explicit support on Win
update CMakeList.txt file to build with AVX defines for Asctenc
ATEEncoder now establishes bc encoding based on version of libate available, or is this statically linked with -late
fix new warnings that popped up with Xcode 12.2

+++ source/etc2comp/EtcBlock4x4Encoding_R11.cpp
@@ -531,12 +531,12 @@ namespace Etc
 			selectorBits |= ((uint64_t)m_redSelectors[uiPixel]) << uiShift;
 		}
 
-		m_pencodingbitsR11->data.selectors0 = selectorBits >> (uint64_t)40;
-		m_pencodingbitsR11->data.selectors1 = selectorBits >> (uint64_t)32;
-		m_pencodingbitsR11->data.selectors2 = selectorBits >> (uint64_t)24;
-		m_pencodingbitsR11->data.selectors3 = selectorBits >> (uint64_t)16;
-		m_pencodingbitsR11->data.selectors4 = selectorBits >> (uint64_t)8;
-		m_pencodingbitsR11->data.selectors5 = selectorBits;        
+		m_pencodingbitsR11->data.selectors0 = uint32_t(selectorBits >> (uint64_t)40);
+		m_pencodingbitsR11->data.selectors1 = uint32_t(selectorBits >> (uint64_t)32);
+		m_pencodingbitsR11->data.selectors2 = uint32_t(selectorBits >> (uint64_t)24);
+		m_pencodingbitsR11->data.selectors3 = uint32_t(selectorBits >> (uint64_t)16);
+		m_pencodingbitsR11->data.selectors4 = uint32_t(selectorBits >> (uint64_t)8);
+		m_pencodingbitsR11->data.selectors5 = uint32_t(selectorBits);
 	}
 
 	// ----------------------------------------------------------------------------------------------------
